+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `withoutMargin` prop on `Modal.Paragraph`.
 - Fix: Replace `Title` component by `TitleWithStroke` on `RewardCardTitle`.
 
 ## [2.76.0] - 2020-07-01